clean code + comments + interface implementation for services
