+++ house-pi/src/main/java/com/rentals/house/dto/LoginRequest.java
@@ -4,9 +4,6 @@ import lombok.Data;
 
 @Data
 public class LoginRequest {
-
   String email;
-
   String password;
-
 }
