+++ house-pi/src/main/java/com/rentals/house/model/Rental.java
@@ -6,7 +6,6 @@ import jakarta.persistence.Entity;
 import lombok.Data;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Data
 @Entity
