+++ house-pi/src/main/java/com/rentals/house/repository/RentalRepository.java
@@ -6,5 +6,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface RentalRepository extends JpaRepository<Rental, Long> {
-
 }
