+++ house-pi/src/main/java/com/rentals/house/repository/UserRepository.java
@@ -9,5 +9,4 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long>  {
   Optional<User> findByEmail(String email);
-  boolean existsByEmail(String email);
 }
