+++ house-pi/src/main/java/com/rentals/house/service/FileStorageService.java
@@ -1,14 +1,6 @@
 package com.rentals.house.service;
 
-import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-
 
 public interface FileStorageService {
   public String storeFile(MultipartFile file);
