+++ house-pi/src/main/java/com/rentals/house/service/JWTService.java
@@ -3,7 +3,6 @@ package com.rentals.house.service;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.jose.jws.MacAlgorithm;
 import org.springframework.security.oauth2.jwt.JwsHeader;
 import org.springframework.security.oauth2.jwt.JwtClaimsSet;
@@ -11,26 +10,6 @@ import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.jwt.JwtEncoderParameters;
 import org.springframework.stereotype.Service;
 
-
-@Service
-public class JWTService {
-
-
-  private final JwtEncoder jwtEncoder;
-
-  public JWTService(JwtEncoder jwtEncoder) {
-    this.jwtEncoder = jwtEncoder;
-  }
-
-  public String generateToken(String email) {
-    Instant now = Instant.now();
-    JwtClaimsSet claims = JwtClaimsSet.builder()
-      .issuer("self")
-      .issuedAt(now)
-      .expiresAt(now.plus(1, ChronoUnit.DAYS))
-      .claim("email", email)
-      .build();
-    JwtEncoderParameters jwtEncoderParameters = JwtEncoderParameters.from(JwsHeader.with(MacAlgorithm.HS256).build(), claims);
-    return this.jwtEncoder.encode(jwtEncoderParameters).getTokenValue();
-  }
+public interface JWTService {
+  public String generateToken(String email);
 }
