+++ house-pi/src/main/java/com/rentals/house/service/impl/FileStorageServiceImpl.java
@@ -1,15 +1,9 @@
 package com.rentals.house.service.impl;
 
 import com.rentals.house.service.FileStorageService;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -18,6 +12,7 @@ import java.nio.file.StandardCopyOption;
 @Service
 public class FileStorageServiceImpl implements FileStorageService {
 
+  // STORE FILES IN THE PICTURES DIRECTORY
   public String storeFile(MultipartFile file) {
     if (file.isEmpty()) {
       throw new RuntimeException("Le fichier est vide.");
